backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
alembic>=1.13.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
google-genai>=0.3.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
firebase-admin>=6.5.0
httpx>=0.27.0
python-multipart>=0.0.9
google-cloud-firestore>=2.15.0
